Added ability to find items "available" for selection, matching or not.
"available" items are either currently selected in the facet group, or
could be selected by the user. The count of available items will likely
be the most useful number to display in the UI to guide the user, and
if no items are available for a facet value that facet value will
likely be hidden.

Potential scenarios that should now be covered:
 - Union in group (across facets) and in each facet
 - Union in group (across facets), intersection in each facet
 - Intersection in group (across facets), union in each facet
 - Intersection in group (across facets) and in each facet